Refactored database compliance (#28)
* :construction: Changed ID type from Integer to Long for DrugReservation

* :white_check_mark: Fixed DrugReservation tests

* :construction: Changed Drug ID type from Integer to Long

* :construction: Changed Pharmacy ID type from Integer to Long

* :construction: Changed SpecialOffer ID type from Integer to Long and changed id fields to concrete object fields

* :construction: Changed EmployedUser ID type from Integer to Long

* :construction: Changed remaining User class ID types from Integer to Long

* :construction::recycle:Refactored EmployedUser into separate Pharmacist and Dermatologist classes

* :construction: Debugging database

* :card_file_box: Implemented Term, EmployedUser and Pharmacy relation

* :memo: Updated documentation

+++ src/main/java/com/schnabel/schnabel/drugreservations/service/IDrugReservationService.java
@@ -8,7 +8,7 @@ import com.schnabel.schnabel.misc.interfaces.ICrudService;
 /**
  * Drug reservation service interface
  */
-public interface IDrugReservationService extends ICrudService<DrugReservation, Integer>
+public interface IDrugReservationService extends ICrudService<DrugReservation, Long>
 {
     /**
      * Get drug reservations valid for <b>date</b>
