+++ src/main/java/com/schnabel/schnabel/drugs/repository/IDrugRepository.java
@@ -4,7 +4,10 @@ import com.schnabel.schnabel.drugs.model.Drug;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface IDrugRepository extends CrudRepository<Drug, Integer>
+/**
+ * Drug repository interface
+ */
+public interface IDrugRepository extends CrudRepository<Drug, Long>
 {
         
 }
