+++ src/main/java/com/schnabel/schnabel/pharmacies/service/ITermService.java
@@ -3,5 +3,9 @@ package com.schnabel.schnabel.pharmacies.service;
 import com.schnabel.schnabel.misc.interfaces.ICrudService;
 import com.schnabel.schnabel.pharmacies.model.Term;
 
-public interface ITermService extends ICrudService<Term, Integer> {
+/**
+ * Term service interface
+ */
+public interface ITermService extends ICrudService<Term, Long>
+{
 }
