+++ src/main/java/com/schnabel/schnabel/pharmacies/service/PharmacyService.java
@@ -7,12 +7,15 @@ import com.schnabel.schnabel.pharmacies.repository.IPharmacyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+/**
+ * Pharmacy service implementation
+ */
 @Service
-public class PharmacyService extends CrudService<Pharmacy, Integer> implements IPharmacyService
+public class PharmacyService extends CrudService<Pharmacy, Long> implements IPharmacyService
 {
     @Autowired
     public PharmacyService(IPharmacyRepository pharmacyRepository)
     {
-		super(pharmacyRepository);
+		  super(pharmacyRepository);
     }
 }
