+++ src/main/java/com/schnabel/schnabel/users/service/IPharmacyAdminService.java
@@ -3,6 +3,6 @@ package com.schnabel.schnabel.users.service;
 import com.schnabel.schnabel.misc.interfaces.ICrudService;
 import com.schnabel.schnabel.users.model.PharmacyAdmin;
 
-public interface IPharmacyAdminService extends ICrudService<PharmacyAdmin, Integer>
+public interface IPharmacyAdminService extends ICrudService<PharmacyAdmin, Long>
 {
 }
